removes test page
